Added telegram bot part.
