+++ .idea/misc.xml
@@ -3,9 +3,7 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.12 (spade-llm)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12 (spade)" project-jdk-type="Python SDK">
-    <output url="file://$PROJECT_DIR$/out" />
-  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.10" project-jdk-type="Python SDK" />
   <component name="PyPackaging">
     <option name="earlyReleasesAsUpgrades" value="true" />
   </component>
