--- spade_llm/demo/platform/tg-example/config.yaml
@@ -1,21 +1,0 @@
-# Configuration of the messaging system. For demo internal in-memory messaging is used
-messaging:
-  type_name: spade_llm.core.messaging.DictionaryMessageService
-# Configuration of the agent state storage. For demo internal in-memory storage is used
-storage:
-  type_name: spade_llm.core.storage.InMemoryStorageFactory
-# Agents to host in the system
-agents:
-  # Echo agent which replies back to the sender of the message with the same content
-  tg-example:
-    #type_name: spade_llm.demo.platform.echo.agents.EchoAgentHandler
-    type_name: spade_llm.demo.platform.tg-example.agents.TelegramAgentHandler
-  # Console agent which listens for commands from console and sends them to the delegate agent
-  tgconsole:
-    type_name: spade_llm.agents.TgChat.TelegramAgent
-    args:
-      # Delegate agent to use
-      bot_token: # Here should be token
-      delegate_type: tg-example
-# Agents to wait for before stopping the system. The system stops when all agents from this list are stopped.
-wait_for_agents: [tgconsole]
