+++ spade_llm/demo/platform/tg_example/config.yaml
@@ -1,0 +1,20 @@
+# Configuration of the messaging system. For demo internal in-memory messaging is used
+messaging:
+  type_name: spade_llm.core.messaging.DictionaryMessageService
+# Configuration of the agent state storage. For demo internal in-memory storage is used
+storage:
+  type_name: spade_llm.core.storage.InMemoryStorageFactory
+# Agents to host in the system
+agents:
+  # Агент отвечающий за ответ на запрос пользователя (В данном случае просто возвращается его же сообщение)
+  echo:
+    type_name: spade_llm.demo.platform.echo.agents.EchoAgentHandler
+  # Бот отвечает за получение и отправку сообщений пользователю, и за инициализацию тг-бота
+  bot:
+    type_name: spade_llm.demo.platform.tg_example.agents.Bot
+    args:
+      bot_token: # here should be TgBot token
+      delegate_type: echo
+# Agents to wait for before stopping the system. The system stops when all agents from this list are stopped.
+wait_for_agents: [ bot ]
+
